add test to request options market data (price)

+++ src/trading/api/contracts/option_contracts.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from ibapi.contract import Contract
 
 
-def get_options_contract(ticker: str, contract_strike: int | str = "",
+def get_options_contract(ticker: str, contract_strike: float | str = "",
                          expiry_date: str | None = None, right: str = "C") -> Contract:
     """Return the option contract object."""
     if expiry_date == "today":
